src/common/requirements.txt
python-decouple==3.3
pymongo==3.11.0
pandas==1.1.3